# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"/"}],"paths":{"/api/schedules/{scheduleId}/queue":{"post":{"tags":["queue-controller"],"operationId":"enter","parameters":[{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"X-User-Id","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnterResponse"}}}}}}},"/api/reservations":{"post":{"tags":["reservation-controller"],"operationId":"hold","parameters":[{"name":"X-User-Id","in":"header","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HoldRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HoldResponse"}}}}}}},"/api/reservations/{reservationId}/confirm":{"post":{"tags":["reservation-controller"],"operationId":"confirm","parameters":[{"name":"X-User-Id","in":"header","required":true,"schema":{"type":"string"}},{"name":"reservationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConfirmResponse"}}}}}}},"/api/schedules/{scheduleId}":{"get":{"tags":["catalog-controller"],"operationId":"schedule","parameters":[{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduleDetailResponse"}}}}}}},"/api/schedules/{scheduleId}/seats/stream":{"get":{"tags":["seat-status-stream-controller"],"operationId":"stream","parameters":[{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/schedules/{scheduleId}/seats/status":{"get":{"tags":["catalog-controller"],"operationId":"status","parameters":[{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SeatStatusResponse"}}}}}}},"/api/schedules/{scheduleId}/seats/layout":{"get":{"tags":["catalog-controller"],"operationId":"layout","parameters":[{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SeatLayoutResponse"}}}}}}},"/api/schedules/{scheduleId}/queue/stream":{"get":{"tags":["queue-stream-controller"],"operationId":"stream_1","parameters":[{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/schedules/{scheduleId}/queue/me":{"get":{"tags":["queue-controller"],"operationId":"me","parameters":[{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"X-User-Id","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}}}},"/api/reservations/{reservationId}":{"get":{"tags":["reservation-controller"],"operationId":"get","parameters":[{"name":"X-User-Id","in":"header","required":true,"schema":{"type":"string"}},{"name":"reservationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}},"delete":{"tags":["reservation-controller"],"operationId":"cancel","parameters":[{"name":"X-User-Id","in":"header","required":true,"schema":{"type":"string"}},{"name":"reservationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/events":{"get":{"tags":["catalog-controller"],"operationId":"events","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventSummaryResponse"}}}}}}}},"/api/events/{eventId}":{"get":{"tags":["catalog-controller"],"operationId":"event","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventDetailResponse"}}}}}}}},"components":{"schemas":{"EnterResponse":{"type":"object","properties":{"position":{"type":"integer","format":"int64"}}},"HoldRequest":{"type":"object","properties":{"scheduleId":{"type":"integer","format":"int64"},"seatId":{"type":"integer","format":"int64"}},"required":["scheduleId","seatId"]},"HoldResponse":{"type":"object","properties":{"reservationId":{"type":"integer","format":"int64"},"expiresAt":{"type":"string","format":"date-time"}}},"ConfirmResponse":{"type":"object","properties":{"reservationId":{"type":"integer","format":"int64"},"status":{"type":"string"},"paymentStatus":{"type":"string"}}},"EventRef":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"venue":{"type":"string"}}},"ScheduleDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"startsAt":{"type":"string","format":"date-time"},"event":{"$ref":"#/components/schemas/EventRef"}}},"SseEmitter":{"type":"object","properties":{"timeout":{"type":"integer","format":"int64"}}},"SeatStatusResponse":{"type":"object","properties":{"scheduleId":{"type":"integer","format":"int64"},"sections":{"type":"array","items":{"$ref":"#/components/schemas/SectionStatusResponse"}}}},"SectionStatusResponse":{"type":"object","properties":{"sectionId":{"type":"integer","format":"int64"},"seatCount":{"type":"integer","format":"int32"},"bitmap":{"type":"string","format":"byte"}}},"SeatLayoutResponse":{"type":"object","properties":{"scheduleId":{"type":"integer","format":"int64"},"sections":{"type":"array","items":{"$ref":"#/components/schemas/SectionResponse"}}}},"SeatResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rowNo":{"type":"integer","format":"int32"},"colNo":{"type":"integer","format":"int32"}}},"SectionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"seatCount":{"type":"integer","format":"int32"},"seats":{"type":"array","items":{"$ref":"#/components/schemas/SeatResponse"}}}},"StatusResponse":{"type":"object","properties":{"position":{"type":"integer","format":"int64"},"admitted":{"type":"boolean"},"token":{"type":"string"}}},"ReservationResponse":{"type":"object","properties":{"reservationId":{"type":"integer","format":"int64"},"scheduleId":{"type":"integer","format":"int64"},"seatId":{"type":"integer","format":"int64"},"status":{"type":"string"},"paymentStatus":{"type":"string"},"paymentTransactionId":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"}}},"EventSummaryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"venue":{"type":"string"},"openAt":{"type":"string","format":"date-time"}}},"EventDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"venue":{"type":"string"},"openAt":{"type":"string","format":"date-time"},"schedules":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"ScheduleResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"startsAt":{"type":"string","format":"date-time"}}}}}}